+++ app/src/main/java/com/example/krani/myapplication/Sziltan/Vektor.java
@@ -1,42 +1,42 @@
 package com.example.krani.myapplication.Sziltan;
 
 public class Vektor {
-    private float x;
-    private float y;
-    private float z;
+    private double x;
+    private double y;
+    private double z;
     public Vektor(){
 
     }
-    public Vektor(float x, float y, float z){
+    public Vektor(double x, double y, double z){
        this.x = x;
        this.y = y;
        this.z = z;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public void setZ(float z) {
+    public void setZ(double z) {
         this.z = z;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public float getZ() {
+    public double getZ() {
         return z;
     }
-    public void setXYZ(float x, float y, float z){
+    public void setXYZ(double x, double y, double z){
         this.z = z;
         this.y = y;
         this.x = x;
